fix(plugins): keep catalog layer independently green
