+++ src/gateway/server-methods-list.test.ts
@@ -133,6 +133,9 @@ describe("listGatewayMethods", () => {
     "sessions.github.status",
     "sessions.github.confirm",
     "sessions.title.prepare",
+    "users.mentionable",
+    "mentions.list",
+    "mentions.dismiss",
     "transcripts.list",
     "transcripts.get",
     "models.authOrderSet",
